Add files via upload
To ensure the correct execution of the Python file, it is essential to first unzip the micasense folder, as it contains the essential functions for radiometric processing and spectral index generation. Additionally, a dedicated environment must be created in Anaconda using the micasense_conda_env.yml file, which automatically installs all the libraries required by the project. If you need additional information about the configuration or associated libraries, you can consult the official MicaSense RedEdge-MX camera guide available in the manufacturer's technical documentation. Once the environment is configured, the script allows you to calculate the descriptive statistics of the different vegetation indices and generate the outputs necessary for computational analysis.
https://micasense.github.io/imageprocessing/MicaSense%20Image%20Processing%20Setup.html